# apollo-openapi.yaml
openapi: 3.0.1
info:
  title: Apollo OpenAPI
  description: |
    <p>Apollo配置中心OpenAPI接口文档</p>

    <h2>认证方式</h2>
    <p>所有 API 接口都需要通过 <code>Authorization</code> header 进行身份验证。</p>

    <h3>获取 Token 的方式：</h3>
    <ol>
      <li><b>Portal 管理界面获取</b>：登录 Portal → 管理员工具 → 开放平台授权管理 → 创建第三方应用，获取 Token。</li>
      <li><b>Token 格式</b>：<code>Authorization: token_value</code></li>
      <li><b>Token 权限</b>：按应用/环境/命名空间授予，建议不同用途分别创建。</li>
    </ol>

    <h3>使用示例</h3>
    <pre><code class="language-bash">curl -X GET "http://localhost:8070/openapi/v1/apps" \
    -H "Authorization: your_token_here"</code></pre>

  version: 1.0.0
security:
  - ApiKeyAuth: []
...
paths:
...
  /openapi/v1/apps/{appId}/namespaces/{namespaceName}/permission-init:
    post:
      summary: 初始化应用权限 (new added)
      operationId: initAppPermission
      deprecated: false
      description: ''
      tags:
        - Permission Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: namespaceName
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: operator
          in: query
          description: '操作人用户名'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: '初始化成功'
  /openapi/v1/apps/{appId}/envs/{env}/clusters/{clusterName}/namespaces/permission-init:
    post:
      summary: 初始化集群Namespace权限 (new added)
      operationId: initClusterNamespacePermission
      deprecated: false
      description: ''
      tags:
        - Permission Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: env
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: clusterName
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: operator
          in: query
          description: '操作人用户名'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ''
  /openapi/v1/apps/{appId}/permissions/{permissionType}:
    get:
      summary: 校验应用权限 (new added)
      operationId: hasAppPermission
      deprecated: false
      description: ''
      tags:
        - Permission Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: permissionType
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: userId
          in: query
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenPermissionConditionDTO'
              example:
                hasPermission: false
  /openapi/v1/apps/{appId}/namespaces/{namespaceName}/permissions/{permissionType}:
    get:
      summary: 校验Namespace权限 (new added)
      operationId: hasNamespacePermission
      deprecated: false
      description: ''
      tags:
        - Permission Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: namespaceName
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: permissionType
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: userId
          in: query
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenPermissionConditionDTO'
              example:
                hasPermission: false
  /openapi/v1/apps/{appId}/envs/{env}/namespaces/{namespaceName}/permissions/{permissionType}:
    get:
      summary: 校验环境Namespace权限 (new added)
      operationId: hasEnvNamespacePermission
      deprecated: false
      description: ''
      tags:
        - Permission Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: env
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: namespaceName
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: permissionType
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: userId
          in: query
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenPermissionConditionDTO'
              example:
                hasPermission: false
  /openapi/v1/apps/{appId}/envs/{env}/clusters/{clusterName}/namespaces/permissions/{permissionType}:
    get:
      summary: 校验集群Namespace权限 (new added)
      operationId: hasClusterNamespacePermission
      deprecated: false
      description: ''
      tags:
        - Permission Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: env
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: clusterName
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: permissionType
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: userId
          in: query
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenPermissionConditionDTO'
              example:
                hasPermission: false
  /openapi/v1/permissions/root:
    get:
      summary: 校验Root权限 (new added)
      operationId: hasRootPermission
      deprecated: false
      description: ''
      tags:
        - Permission Management
      parameters:
        - name: userId
          in: query
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenPermissionConditionDTO'
              example:
                hasPermission: false
  /openapi/v1/apps/{appId}/envs/{env}/namespaces/{namespaceName}/role-users:
    get:
      summary: 获取环境Namespace用户 (new added)
      operationId: getNamespaceEnvRoleUsers
      deprecated: false
      description: ''
      tags:
        - Permission Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: env
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: namespaceName
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenEnvNamespaceRoleUserDTO'
                description:
              example:
                appId: ''
                namespaceName: ''
                modifyRoleUsers:
                  - userId: ''
                    name: ''
                    email: ''
                    enabled: 0
                releaseRoleUsers:
                  - userId: ''
                    name: ''
                    email: ''
                    enabled: 0
                env: ''
  /openapi/v1/apps/{appId}/envs/{env}/namespaces/{namespaceName}/roles/{roleType}:
    post:
      summary: 分配环境Namespace角色 (new added)
      operationId: assignNamespaceEnvRoleToUser
      deprecated: false
      description: ''
      tags:
        - Permission Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: env
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: namespaceName
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: roleType
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: operator
          in: query
          description: '操作人用户名'
          required: false
          schema:
            type: string
        - name: userId
          in: query
          description: 'the user which role assigned to'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
    delete:
      summary: 移除环境Namespace角色 (new added)
      operationId: removeNamespaceEnvRoleFromUser
      deprecated: false
      description: ''
      tags:
        - Permission Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: env
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: namespaceName
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: roleType
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: operator
          in: query
          description: '操作人用户名'
          required: false
          schema:
            type: string
        - name: userId
          in: query
          description: 'the user which role removed from'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
...
  /openapi/v1/apps/{appId}/namespaces/{namespaceName}/role-users:
    get:
      summary: 获取Namespace用户 (new added)
      operationId: getNamespaceRoles
      deprecated: false
      description: ''
      tags:
        - Permission Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: namespaceName
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenNamespaceRoleUserDTO'
                description:
              example:
                appId: ''
                namespaceName: ''
                modifyRoleUsers:
                  - userId: ''
                    name: ''
                    email: ''
                    enabled: 0
                releaseRoleUsers:
                  - userId: ''
                    name: ''
                    email: ''
                    enabled: 0
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Apollo OpenAPI访问令牌
        **获取方式：**
        1. 登录Apollo Portal管理界面
        2. 进入"管理员工具" -> "开放平台授权管理"
        3. 创建第三方应用，获取Token
        **使用格式：**
        ```
        Authorization: your_token_here
        ```
        **权限说明：**
        - Token具有对应应用的读写权限
        - 不同Token可能有不同的环境和命名空间权限
        - 建议为不同用途创建不同的Token
  schemas:
...
    OpenPermissionConditionDTO:
      type: object
      properties:
        hasPermission:
          type: boolean
          description: ''
...
    OpenEnvNamespaceRoleUserDTO:
      type: object
      properties:
        appId:
          type: string
          description: ''
        namespaceName:
          type: string
          description: ''
        modifyRoleUsers:
          type: array
          items:
            $ref: '#/components/schemas/OpenUserInfoDTO'
            description:
          description: ''
        releaseRoleUsers:
          type: array
          items:
            $ref: '#/components/schemas/OpenUserInfoDTO'
            description:
          description: ''
        env:
          type: string
          description: ''
...
servers: []